Examples/Tutorials/Flask Tutorials/Flask_Blog/requirements.txt
attrs==19.3.0
bcrypt==3.1.7
binarytree==4.1.0
blinker==1.4
cffi==1.14.0
click==7.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.1.0
packaging==20.1
Pillow==7.1.1
pluggy==0.13.1
py==1.8.1
pycparser==2.20
pyparsing==2.4.6
pytest==5.3.4
six==1.14.0
SQLAlchemy==1.3.16
wcwidth==0.1.8
Werkzeug==1.0.1
WTForms==2.2.1
zipp==2.0.1
